Added application argument protection. The crash monitor will only run when the command line argument --enableCrashMonitor is present. This should close issue #94 .

+++ source/application/AMCrashMonitorSupport.cpp
@@ -73,6 +73,10 @@ void AMCrashMonitorSupport::monitor(){
 
 	QCoreApplication *app = QApplication::instance();
 
+	QStringList appArguments = app->arguments();
+	if(!appArguments.contains("--enableCrashMonitor"))
+		return;
+
 	QStringList arguments;
 	arguments << "-m";
 	arguments << app->applicationFilePath();
@@ -83,6 +87,12 @@ void AMCrashMonitorSupport::monitor(){
 }
 
 void AMCrashMonitorSupport::report(){
+	QCoreApplication *app = QApplication::instance();
+
+	QStringList appArguments = app->arguments();
+	if(!appArguments.contains("--enableCrashMonitor"))
+		return;
+
 	kill(globalCrashMonitorPID, SIGUSR2);
 }
